requirements.txt
fireworks-client
google-cloud-vision
pytesseract
Pillow